Hide the Menu button when there's nothing to show
Note that the `content_block` helper isn't available when we call the
`Layout::SubnavigationComponent.new.render?` and
`Layout::TopLinksComponent.new.render?` methods. As a workaround, we're
duplicating what the `content_block` helper does. Another options would
be creating a component which does what this helper does, like we did
with the `Shared::LinkListComponent`.